remove all the duplicated code in iOS and android demos
